Move out ExecutedPayload and ExecutionInfo
