+++ crates/op-rbuilder/src/primitives/reth/execution.rs
@@ -1,0 +1,80 @@
+//! [Source]()
+use reth_node_api::NodePrimitives;
+use alloy_primitives::{Address, TxHash, B256, U256};
+use std::collections::HashSet;
+use alloy_consensus::Transaction;
+use alloy_primitives::private::alloy_rlp::Encodable;
+
+/// Holds the state after execution
+#[derive(Debug)]
+pub struct ExecutedPayload<N: NodePrimitives> {
+    /// Tracked execution info
+    pub info: ExecutionInfo<N>,
+    /// Withdrawal hash.
+    pub withdrawals_root: Option<B256>,
+}
+
+#[derive(Default, Debug)]
+pub struct ExecutionInfo<N: NodePrimitives> {
+    /// All executed transactions (unrecovered).
+    pub executed_transactions: Vec<N::SignedTx>,
+    /// The recovered senders for the executed transactions.
+    pub executed_senders: Vec<Address>,
+    /// The transaction receipts
+    pub receipts: Vec<N::Receipt>,
+    /// All gas used so far
+    pub cumulative_gas_used: u64,
+    /// Estimated DA size
+    pub cumulative_da_bytes_used: u64,
+    /// Tracks fees from executed mempool transactions
+    pub total_fees: U256,
+    /// Tracks the reverted transaction hashes to remove from the transaction pool
+    pub reverted_tx_hashes: HashSet<TxHash>,
+    #[cfg(feature = "flashblocks")]
+    /// Index of the last consumed flashblock
+    pub last_flashblock_index: usize,
+}
+
+
+impl<N: NodePrimitives> ExecutionInfo<N> {
+    /// Create a new instance with allocated slots.
+    pub fn with_capacity(capacity: usize) -> Self {
+        Self {
+            executed_transactions: Vec::with_capacity(capacity),
+            executed_senders: Vec::with_capacity(capacity),
+            receipts: Vec::with_capacity(capacity),
+            cumulative_gas_used: 0,
+            cumulative_da_bytes_used: 0,
+            total_fees: U256::ZERO,
+            reverted_tx_hashes: HashSet::new(),
+            #[cfg(feature = "flashblocks")]
+            last_flashblock_index: 0,
+        }
+    }
+
+    /// Returns true if the transaction would exceed the block limits:
+    /// - block gas limit: ensures the transaction still fits into the block.
+    /// - tx DA limit: if configured, ensures the tx does not exceed the maximum allowed DA limit
+    ///   per tx.
+    /// - block DA limit: if configured, ensures the transaction's DA size does not exceed the
+    ///   maximum allowed DA limit per block.
+    pub fn is_tx_over_limits(
+        &self,
+        tx: &N::SignedTx,
+        block_gas_limit: u64,
+        tx_data_limit: Option<u64>,
+        block_data_limit: Option<u64>,
+    ) -> bool {
+        if tx_data_limit.is_some_and(|da_limit| tx.length() as u64 > da_limit) {
+            return true;
+        }
+
+        if block_data_limit
+            .is_some_and(|da_limit| self.cumulative_da_bytes_used + (tx.length() as u64) > da_limit)
+        {
+            return true;
+        }
+
+        self.cumulative_gas_used + tx.gas_limit() > block_gas_limit
+    }
+}
